LPS-182960 inline GraphUtil since it is not used by any other classes
